+++ component/action/SortAction.js
@@ -31,7 +31,7 @@ module.exports = class SortAction extends Base {
             return this.controller.renderMeta(this.template, this.controller.getMetaParams({sortAttr}));
         }
         const data = this.validateData(this.getPostParam('order'));
-        for (const config of view.getBehaviorsByClassAndAttr(SortOrderBehavior, attrName)) {
+        for (const config of view.behaviors.getAllByClassAndAttr(SortOrderBehavior, attrName)) {
             const behavior = this.spawn(config, {
                 owner: this,
                 user: this.user