+++ message/app/ru.js
@@ -1,3 +1,9 @@
+/**
+ * @copyright Copyright (c) 2020 Maxim Khorin <maksimovichu@gmail.com>
+ */
+'use strict';
+
 module.exports = {
+
     '<span class="no-access">[no access]</span>': '<span class="no-access">[нет доступа]</span>',
 };