+++ web/message/ru.js
@@ -6,11 +6,12 @@
 Object.assign(Jam.I18n.defaults, {
 
     'Delete relation sorting': 'Удалить сортировку отношения',
-    'Delete relation sorting?': 'Удалить сортировку отношения?',
+    'Digital signature created': 'Цифровая подпись создана',
 
     'Select class': 'Выберите класс',
+    'Sign data': 'Подписать данные',
     'Sort related objects': 'Сортировать связанные объекты',
     'Source data': 'Исходные данные',
 
-    'Transition in progress...': 'Состояние в процессе перехода...'
+    'Transition in progress...': 'Выполняется переход...'
 });